refactor: move shared enums to enums.py
