+++ google/cloud/alloydb/connector/async_connector.py
@@ -24,7 +24,7 @@ import google.auth.transport.requests
 
 import google.cloud.alloydb.connector.asyncpg as asyncpg
 from google.cloud.alloydb.connector.client import AlloyDBClient
-from google.cloud.alloydb.connector.instance import IPTypes
+from google.cloud.alloydb.connector.enums import IPTypes
 from google.cloud.alloydb.connector.instance import RefreshAheadCache
 from google.cloud.alloydb.connector.utils import generate_keys
