+++ google/cloud/alloydb/connector/connector.py
@@ -26,7 +26,7 @@ from google.auth import default
 from google.auth.credentials import with_scopes_if_required
 
 from google.cloud.alloydb.connector.client import AlloyDBClient
-from google.cloud.alloydb.connector.instance import IPTypes
+from google.cloud.alloydb.connector.enums import IPTypes
 from google.cloud.alloydb.connector.instance import RefreshAheadCache
 import google.cloud.alloydb.connector.pg8000 as pg8000
 from google.cloud.alloydb.connector.utils import generate_keys
